\name{cfwlab-package}
\alias{cfwlab-package}
\alias{cfwlab}
\docType{package}
\title{CFW genome-wide mapping study data}
\description{Description goes here.}
\section{Introduction}{
  Introduction goes here.
}
\author{Peter Carbonetto and John Novembre}
\seealso{
  \url{https://github.com/pcarbo/cfwlab}
}
