\name{cfwlab-package}
\alias{cfwlab-package}
\alias{cfwlab}
\docType{package}
\title{CFW genome-wide association study data}
\description{Description goes here.}
\section{Introduction}{
  Introduction goes here.
}
\references{
  
  Parker, C. C., Gopalakrishnan, S., Carbonetto, P., Gonzales, N.M.,
  Leung, E., Park, Y.J., Aryee, E., Davis, J., Blizard, D.A.,
  Ackert-Bicknell, C.L., Lionikas, A., Pritchard, J.K. and Palmer,
  A.A. (2016) Genome-wide association study of behavioral, physiological
  and gene expression traits in outbred CFW mice. \emph{Nature Genetics}
  \bold{48}, 919--926. \url{http://dx.doi.org/10.1038/ng.3609}

  Parker, C. C., Gopalakrishnan, S., Carbonetto, P., Gonzales, N.M.,
  Leung, E., Park, Y.J., Aryee, E., Davis, J., Blizard, D.A.,
  Ackert-Bicknell, C.L., Lionikas, A., Pritchard, J.K. and Palmer,
  A.A. (2016) Data from: Genome-wide association study of behavioral,
  physiological and gene expression traits in outbred CFW
  mice. \emph{Dryad Digital Repository}. 
  \url{http://dx.doi.org/10.5061/dryad.2rs41}
}
\author{Peter Carbonetto and John Novembre}
\seealso{
  \url{https://github.com/pcarbo/cfwlab}
}
